Rename type aliases
